+++ gauss_apps/libraries/inverse_app.py
@@ -106,7 +106,7 @@ start_input = VBox( [start_text, start_btn])
 
 # Create a HTML object to show the augmented equation in LaTeX format.
 amat_latex = HTMLMath(
-  value=r'The augmented matrix $(\bf{A}|\bf{B})$ will appear here.',
+  value=r'The augmented matrix $(\bf{A}|\bf{I})$ will appear here.',
   layout=Layout(justify_content='center'))
 
 # # Put together all inputs.
@@ -238,8 +238,8 @@ def update_aug_matrix(change):
 
   # Create the string with the initial augmented matrix in LaTeX format.
   latex_str = r'The augmented matrix $(\bf{A}|\bf{I})$ is:' + '\n $$' \
-    + convert_mat_to_latex(M, color_dict, 'rrr|rrr') + r'$$ with ' \
-    + detA_str + '.'
+    + r'\left [' + convert_mat_to_latex(M, color_dict, 'rrr|rrr')     \
+    + r'\right ] $$ with ' + detA_str + '.'
  
   # Update object "amat_latex" with the augmented matrix in LaTeX format. 
   amat_latex.value=latex_str
@@ -316,8 +316,9 @@ def apply_row_swap(change):
     M_str     = convert_mat_to_latex(M, color_M, 'rrr|rrr')
 
     # Update output.
-    results_html.value += err_str + '$$ $$' + '$$' + M_old_str + row_op_str + \
-      M_str + '$$'
+    results_html.value += err_str + '$$ $$'                     \
+      + '$$ \left[' + M_old_str + r'\right ]'     + row_op_str  \
+      + '\left ['   + M_str     + r'\right ] $$'
 
   # End if.
 
@@ -389,8 +390,9 @@ def apply_row_times_scalar(change):
     M_str     = convert_mat_to_latex(M, color_M, 'rrr|rrr')
 
     # Update output.
-    results_html.value += err_str + '$$ $$' + '$$' + M_old_str + row_op_str \
-      + M_str + '$$'
+    results_html.value += err_str + '$$ $$'                     \
+      + '$$ \left[' + M_old_str + r'\right ]'     + row_op_str  \
+      + '\left ['   + M_str     + r'\right ] $$'
 
   # End if.
 
@@ -468,14 +470,13 @@ def apply_row_add(change):
     M_str     = convert_mat_to_latex(M, color_M, 'rrr|rrr')
 
     # Update output.
-    results_html.value += err_str + '$$ $$' + '$$' + M_old_str + row_op_str \
-      + M_str + '$$'
+    results_html.value += err_str + '$$ $$'                     \
+      + '$$ \left[' + M_old_str + r'\right ]'     + row_op_str  \
+      + '\left ['   + M_str     + r'\right ] $$'
 
   # End if.
 
 #-------------------------------------------------------------------------------
-
-
 
 
 #*******************************************************************************