+++ gauss_apps/libraries/lib_gauss.py
@@ -75,6 +75,57 @@ def row_swap(M, rid1, rid2):
 
 
 #*******************************************************************************
+def col_swap(M, cid1, cid2):
+#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+  '''
+  Swap two columns of a SymPy matrix.
+
+  M:    SymPy matrix to be modified.
+  cid1: integer identificator of a column. One based (1, 2, ...).
+  cid2: identificator of another column. One based (1, 2, ...).
+
+  Returns a new matrix with swapped rows.
+  '''
+#*******************************************************************************
+ 
+  ncols = M.cols            # Get the number of columns.
+
+  if (cid1 == cid2):        # Verify if the swap is not valid
+  
+    print('Error. No columns were swapped.')
+
+    new_M = M.copy()
+
+  elif (cid1 > ncols) or (cid2 > ncols):
+
+    print(f'Error. Please use col. indices from 1 through {nrows}.')
+
+    new_M = M.copy()
+
+  elif (cid1 == 0) or (cid2 == 0):
+  
+    print(f'Error. Please use row indices from 1 through {nrows}.')
+
+    new_M = M.copy()
+
+  else:
+
+    # Create a list of indices.
+    id_lst = [k for k in range(ncols)]    # Create a list of indices.
+
+    id_lst[cid1 - 1] = cid2 - 1           # Swap indices in the list.
+    id_lst[cid2 - 1] = cid1 - 1
+
+    new_M = M[:,id_lst]                   # Reorganize matrix.
+
+  # End if. 
+
+  return new_M
+ 
+#-------------------------------------------------------------------------------
+
+
+#*******************************************************************************
 def row_add(M, new_rid, c1, rid1, c2, rid2):
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
   '''
@@ -109,6 +160,41 @@ def row_add(M, new_rid, c1, rid1, c2, rid2):
 
 
 #*******************************************************************************
+def col_add(M, new_cid, c1, cid1, c2, cid2):
+#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+  '''
+  Set new values for a column that are the linear combination of two existing
+  columns.
+
+  M:        SymPy matrix to be modified.
+  new_cid:  integer identificator of the column to be modified. One based (1, 2,
+            ...).
+  c1:       real, scalar value to multiply column 1.
+  cid1:     integer identificator of column 1. One based (1, 2, ...).
+  c2:       real, scalar value to multiply column 2.
+  cid2:     integer identificator of column 2. One based (1, 2, ...).
+  '''
+#*******************************************************************************
+
+  new_M = M.copy()
+
+  if (new_cid != cid1) and (new_cid != cid2):
+
+    print(f'Error. The index of the new column is {new_cid}.')
+    print(f'It must be either {cid1} or {cid2}.')
+
+  else:
+ 
+    new_M[:, new_cid - 1] = c1*M[:, cid1 - 1] + c2*M[:, cid2 - 1]
+
+  # End if.
+
+  return new_M
+
+#-------------------------------------------------------------------------------
+
+
+#*******************************************************************************
 def row_times_scalar(M, c, rid):
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
   '''
@@ -123,6 +209,27 @@ def row_times_scalar(M, c, rid):
   new_M = M.copy()
 
   new_M[rid - 1, :] = c*M[rid - 1, :]
+
+  return new_M
+
+#-------------------------------------------------------------------------------
+
+
+#*******************************************************************************
+def col_times_scalar(M, c, cid):
+#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+  '''
+  Multiplies the column of a Sympy matrix times a scalar value.
+
+  M:        SymPy matrix to be modified.
+  c:        real, scalar value to multiply the row.
+  cid:      integer identificator of the column. One based (1, 2, ...).
+  '''
+#*******************************************************************************
+
+  new_M = M.copy()
+
+  new_M[:, cid - 1] = c*M[:, cid - 1]
 
   return new_M
 
@@ -182,7 +289,7 @@ def generate_latex_string_per_row(row, color_str=''):
 
 
 #*******************************************************************************
-def convert_mat_to_latex(M, color_dict, align_opt=""):
+def convert_mat_to_latex(M, color_dict={0:'',1:'',2:''}, align_opt=""):
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
   '''
   Convert a Sympy matrix to a LaTeX. It generates the code to print the
@@ -211,7 +318,7 @@ def convert_mat_to_latex(M, color_dict, align_opt=""):
 
   nrows = M.rows
  
-  latex_str = r'\left[ \begin{array}' + r'{{ {:s} }}'.format(align_opt) + '\n'
+  latex_str = r'\begin{array}' + r'{{ {:s} }}'.format(align_opt) + '\n'
 
   for k in range(nrows):
 
@@ -233,7 +340,7 @@ def convert_mat_to_latex(M, color_dict, align_opt=""):
 
   # End for. Loop over rows of the matrix. 
 
-  latex_str += r'\end{array} \right]'
+  latex_str += r'\end{array}'
 
   return latex_str